+++ prompts/prompts.go
@@ -338,6 +338,42 @@ func ResourceLabel(defaultValue string, autoSelect bool) (string, error) {
 	return p.Run()
 }
 
+// TeamName prompts the user to enter a new Team name
+func TeamName(defaultValue string, autoSelect bool) (string, error) {
+	validate := func(input string) error {
+		if len(input) == 0 {
+			return promptui.NewValidationError("Please provide a valid team name")
+		}
+
+		l := manifold.Name(input)
+		if err := l.Validate(nil); err != nil {
+			return promptui.NewValidationError("Please provide a valid team name")
+		}
+
+		return nil
+	}
+
+	label := "Team Name"
+
+	if autoSelect {
+		err := validate(defaultValue)
+		if err != nil {
+			fmt.Println(promptui.FailedValue(label, defaultValue))
+		} else {
+			fmt.Println(promptui.SuccessfulValue(label, defaultValue))
+		}
+		return defaultValue, err
+	}
+
+	p := promptui.Prompt{
+		Label:    label,
+		Default:  defaultValue,
+		Validate: validate,
+	}
+
+	return p.Run()
+}
+
 // Email prompts the user to provide an email *or* accepted the default
 // email value
 func Email(defaultValue string) (string, error) {