+++ cmd/validate.go
@@ -9,16 +9,23 @@ import (
 	"github.com/manifoldco/manifold-cli/errs"
 )
 
-func validateName(cliCtx *cli.Context, option string) (string, error) {
+func validateName(cliCtx *cli.Context, option string, typeName ...string) (string, error) {
 	val := cliCtx.String(option)
 	if val == "" {
 		return val, nil
 	}
 
+	if len(typeName) > 1 {
+		panic("Only one typeName can be provided")
+	} else if len(typeName) == 0 {
+		typeName = make([]string, 1)
+		typeName[0] = option
+	}
+
 	name := manifold.Name(val)
 	if err := name.Validate(nil); err != nil {
 		return "", errs.NewUsageExitError(cliCtx, cli.NewExitError(
-			fmt.Sprintf("You've provided an invalid %s name!", option), -1,
+			fmt.Sprintf("You've provided an invalid %s name!", typeName[0]), -1,
 		))
 	}