+++ prompts/prompts.go
@@ -215,8 +215,9 @@ func SelectRegion(regions []*cModels.Region) (int, string, error) {
 	return prompt.Run()
 }
 
-// SelectTeam prompts the user to select a team from the given list
-func SelectTeam(teams []*iModels.Team, label string) (int, string, error) {
+// SelectTeam prompts the user to select a team from the given list. -1 as the first return value
+// indicates no team has been selected
+func SelectTeam(teams []*iModels.Team, label string, includeNoTeam bool) (int, string, error) {
 	line := func(t *iModels.Team) string {
 		return fmt.Sprintf("%s (%s)", t.Body.Name, t.Body.Label)
 	}
@@ -243,9 +244,13 @@ func SelectTeam(teams []*iModels.Team, label string) (int, string, error) {
 		return idx, label, nil
 	}
 
-	labels := make([]string, len(teams))
+	labels := make([]string, len(teams)+1)
 	for i, t := range teams {
 		labels[i] = line(t)
+	}
+
+	if includeNoTeam {
+		labels = append([]string{"Don't use a team"}, labels...)
 	}
 
 	prompt := promptui.Select{
@@ -253,7 +258,13 @@ func SelectTeam(teams []*iModels.Team, label string) (int, string, error) {
 		Items: labels,
 	}
 
-	return prompt.Run()
+	teamIdx, name, err := prompt.Run()
+
+	if includeNoTeam {
+		return teamIdx - 1, name, err
+	}
+
+	return teamIdx, name, err
 }
 
 // SelectCreateAppName prompts the user to select or create an application