+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Add a newline after signup output, so it won't mess up the prompt.
+- Credit Card number can be between 12 and 19 digits and cvv either 3 or 4.
 
 ## [0.2.6] - 2017-08-16